app/requirements.txt
# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23

# API & Serialization
Flask-CORS==4.0.0
marshmallow==3.20.1

# Authentication & Security
PyJWT==2.8.0
Flask-Limiter==3.5.0

# Caching
Flask-Caching==2.1.0
redis==5.0.1

# Monitoring
prometheus-client==0.19.0

# Production Server
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
